# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.User:
    properties:
      company_id:
        type: string
      email:
        type: string
      id:
        type: string
      is_confirmed:
        type: boolean
      name:
        type: string
      password:
        type: string
      session:
        $ref: '#/definitions/domain.UserSession'
    type: object
  domain.UserSession:
    properties:
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
...
  http.response:
    properties:
      message:
        type: string
    type: object
...
  http.userRefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  http.userRefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  http.userSignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  http.userSignInResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  http.userSignUpRequest:
    properties:
      company_id:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  http.userSignUpResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  http.userUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  title: Syncrepair API
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Delete user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Delete
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.userUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Update
      tags:
      - users
  /users/confirm/{id}:
    post:
      consumes:
      - application/json
      description: Confirm user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: Confirm
      tags:
      - users
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Refresh user tokens
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.userRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.userRefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: Refresh tokens
      tags:
      - users
  /users/sign-in:
    post:
      consumes:
      - application/json
      description: Sign in to user account
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.userSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.userSignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: Sign in
      tags:
      - users
  /users/sign-up:
    post:
      consumes:
      - application/json
      description: Create new user account
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.userSignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.userSignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: Sign up
      tags:
      - users
securityDefinitions:
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"